backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic-settings==2.2.1
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4
PyMuPDF==1.24.3
pytesseract==0.3.10
Pillow==10.3.0
httpx==0.27.0
python-magic==0.4.27
